examples/hikari_tanjun_base/requirements.txt
hikari-tanjun
